# Performance improvement for function dirwalk.