### Issue
https://github.com/masahiro-kasatani/pacvim/issues/2

### Done
#### Add test code for dirwalk
- [commit](https://github.com/masahiro-kasatani/pacvim/commit/a07ed88f4251e0c6c796a155b43fc3b3a27b2bdd)
#### Change from regexp to strings
- [commit](https://github.com/masahiro-kasatani/pacvim/commit/3658180c99020fc6a275d971040942564736ffd1)
#### Run Benchmark

```go
func Benchmark_dirwalk(b *testing.B) {
	for i := 0; i < 100; i++ {
		_, _ = dirwalk(stageDir)
	}
}
```
- Before
```sh
% go test -test.bench=Benchmark_dirwalk -run=^$
goos: darwin
goarch: amd64
pkg: pacvim
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
Benchmark_dirwalk-8   	1000000000	         0.007149 ns/op
PASS
ok  	pacvim	0.424s
```
- After
```sh
% go test -test.bench=Benchmark_dirwalk -run=^$
goos: darwin
goarch: amd64
pkg: pacvim
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
Benchmark_dirwalk-8   	1000000000	         0.006761 ns/op
PASS
ok  	pacvim	0.267s
```